Changing the condition to show the history button to use the revision property of the page to save a query on every page load. See #242.
Also made some slight template modfications.

There still will be times when there is only 1 revision of the page and the history button will display. For example, if after the page is created and saved, it is edited and saved again (not as a new revision). That increments the revision count (because it was revised), but does not create another entry in the revisions table. Or, when the admin uses the 'fix' or 'clean up the database' functions - those can remove all other revisions, leaving only one.

These conditions will be addressed in other tickets

git-svn-id: https://subversion.assembla.com/svn/impresscmsaddons/modules/wiwimod/tasks/242_revision_feature@25348 7a4d4d67-0c3a-0410-b24c-89d50c169223